Refactor names Sprint 1- SocialMeli
