+++ src/main/java/com/meli/socialmeli/controllers/UserController.java
@@ -1,4 +1,4 @@
-package com.meli.socialmeli.controller;
+package com.meli.socialmeli.controllers;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
