+++ src/main/java/com/meli/socialmeli/exceptions/custom/BadRequest.java
@@ -1,4 +1,4 @@
-package com.meli.socialmeli.exception.custom;
+package com.meli.socialmeli.exceptions.custom;
 
 public class BadRequest extends RuntimeException{
     public BadRequest(String message) {
