+++ src/main/java/com/meli/socialmeli/exceptions/custom/NotFoundException.java
@@ -1,4 +1,4 @@
-package com.meli.socialmeli.exception.custom;
+package com.meli.socialmeli.exceptions.custom;
 
 public class NotFoundException extends  RuntimeException{
     public NotFoundException(String message) {
