+++ src/main/java/com/meli/socialmeli/repositories/IProductRepository.java
@@ -1,0 +1,4 @@
+package com.meli.socialmeli.repositories;
+
+public interface IProductRepository {
+}
