+++ src/main/java/com/meli/socialmeli/repositories/IUserRepository.java
@@ -1,0 +1,4 @@
+package com.meli.socialmeli.repositories;
+
+public interface IUserRepository {
+}
