--- src/main/java/com/meli/socialmeli/repository/IProductRepository.java
@@ -1,4 +1,0 @@
-package com.meli.socialmeli.repository;
-
-public interface IProductRepository {
-}
