--- src/main/java/com/meli/socialmeli/repository/IUserRepository.java
@@ -1,4 +1,0 @@
-package com.meli.socialmeli.repository;
-
-public interface IUserRepository {
-}
