--- src/main/java/com/meli/socialmeli/service/IUserService.java
@@ -1,4 +1,0 @@
-package com.meli.socialmeli.service;
-
-public interface IUserService {
-}
