+++ src/main/java/com/meli/socialmeli/services/IUserService.java
@@ -1,0 +1,4 @@
+package com.meli.socialmeli.services;
+
+public interface IUserService {
+}
